% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide_happy_hour}
\alias{decide_happy_hour}
\title{Find the possibilities of dates}
\usage{
decide_happy_hour(x)
}
\arguments{
\item{x}{ABC}
}
\value{
It will provide possibles Dates
}
\description{
Find the possibilities of dates
}
\examples{
1 + 2
}
